Created a test class that will run through the compression suite once for a single file, testing the compression libraries and the database performance.
Had to remove preparedStatements from DBAdapter because apparently the table name can't be a parameter, and that's a dealbreaker.

Created a stringify function in Main to convert SHA256 digest into a string of hex chars.

CompressedSize and origSize are now doubles instead of integers.  It would probably be better to store the file size in kb, right?